requirements.txt
--find-links https://download.pytorch.org/whl/cu102/torch_stable.html
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
pyannote.audio>=2.1.1
whisper-timestamped
ffmpeg-python
python-dotenv
transformers